[UIMA-1866] change CVD to give overwrite confirmation dialog when writing out files.
git-svn-id: https://svn.apache.org/repos/asf/uima/uimaj/trunk@1165799 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-tools/src/main/java/org/apache/uima/tools/cvd/control/XCASSaveHandler.java
@@ -55,6 +55,9 @@ public class XCASSaveHandler implements ActionListener {
     int rc = fileChooser.showSaveDialog(this.main);
     if (rc == JFileChooser.APPROVE_OPTION) {
       File xcasFile = fileChooser.getSelectedFile();
+      if (!this.main.confirmOverwrite(xcasFile)) {
+        return;
+      }
       this.main.setXcasFileOpenDir(xcasFile.getParentFile());
       try {
         OutputStream outStream = new BufferedOutputStream(new FileOutputStream(xcasFile));
